Slideshow caching
A more elegant solution to my 'off-timing' slideshow problem...  Discovered IIFEs in JS and tried coding a pre-loader IIFE to cache my slideshow images in an array and then calling a function to run through the slideshow.

Supplemented this by putting the images in list elements in the HTML and adjusting the CSS a bit. Seems to work! :raised_hands: