More modal mods :100:
The modal layout became an issue when viewing on a landscape-oriented mobile device. Fixed this by adjusting the grid template to include columns for the project image and description. Added a media query to detect the change in orientation.

Also resolved the "I can still scroll while the modal is showing!" problem by following this helpful article: https://css-tricks.com/prevent-page-scrolling-when-a-modal-is-open/. Essentially, when the modal is showing: fixed positioning is applied to the body with the top property being set to the current scrollY value, and these values are reset to their initial values when the modal is closed.